# docs/api/external/operations.openapi.yaml
openapi: 3.0.3
info:
  title: MAIster External Operations API
  version: "1.0.0"
  description: |
    Token-authenticated external surface for MAIster (M16). All routes are
    under `/api/v1/ext/` and require a bearer token issued via the
    token-management routes in `docs/api/web.openapi.yaml`. Tokens may be
    project-bound (`project_id` set) or global personal user tokens
    (`project_id = NULL`, Implemented by 0063).

    **Security model.** Every call is authenticated by the `projectToken`
    bearer scheme. The token is 256-bit random (`mai_` + base64url(32 bytes)),
    stored as sha256 at rest, looked up by prefix, and compared with
    `timingSafeEqual`. Token failure kinds map to 401. Project-bound tokens
    must match the addressed project. Global personal user tokens derive the
    target project from URL/server-state, load the owner account, require the
    owner to be active and password-clear, and authorize through current
    project membership/global admin. Cross-project or inaccessible-project
    access hides existence (404). Every route enforces its scope label unless
    the token holds `*`; `*` does not satisfy exact-only
    `hitl:respond:human`.

    **Audit.** Every call writes a `token_audit_log` row. The gate-report
    endpoint writes the success audit row inside the same database transaction
    as the gate update and artifact; failure-path audit rows are written after
    and do not mutate gate or artifact state.

...

    See also:
...

...
servers:
...
security:
  - projectToken: []
...
paths:
...
  /api/v1/ext/projects/{slug}/tasks/{taskId}/triage:
    post:
      summary: (M34 — Implemented; ADR-112 flag/enqueue — Implemented) Submit a triage verdict
      description: |
        Scope `tasks:triage`. The triager's verdict op — ONE transactional
        update of the task's launch-verdict fields plus the triage-status
        stamp. At least one body field is required; every provided field is
        validated against server state (`flowId` ∈ project flows, `runnerId`
        ∈ enabled runner catalog, `promotionMode` ∈ enum, `targetBranch` =
        git ref-name shape). Records a `triage_set` activity entry carrying
        the payload. Writes a `token_audit_log` row with scope `tasks:triage`;
        agent tokens record `actor_label = agent:<id>` and the activity
        actor is the polymorphic `agent` pair (ADR-083).

        **Verdict path (default).** A verdict body (`flowId` / `runnerId` /
        `baseBranch` / `targetBranch` / `promotionMode`) sets
        `tasks.triage_status = 'triaged'`; the task becomes launchable.

        **(ADR-112, Implemented) `flag` — hand to a human.** When `flag: true`,
        the task is marked `tasks.triage_status = 'flagged'` (held — NOT
        launchable; the board shows a "needs review" chip). `flag` is
        **mutually exclusive** with the verdict fields (`flowId` / `runnerId` /
        `baseBranch` / `targetBranch` / `promotionMode`): sending `flag`
        alongside any verdict field is refused 422 `CONFIG`. `flag: true` alone
        is a valid body (the "≥1 field" rule is still met).

        **(ADR-112, Implemented) `enqueue` — auto-launch after triage.** When
        `enqueue: true`, the task's enqueue intent is set
        (`tasks.launch_mode = 'auto'`) so the `auto_launch_triaged` scheduler
        tick launches it once it is launchable (dependencies cleared, global
        cap permitting). `enqueue` is valid ONLY alongside a verdict that
        yields a `flowId` (an existing `tasks.flow_id` or a `flowId` in this
        body); `enqueue` with no resolvable flow, or together with `flag`, is
        refused 422 `CONFIG`. The triager never launches directly — the tick
        performs the launch through the standard precondition choke point.

        **Write-side launchability re-validation (ADR-112, Implemented).** When a
        verdict carries a `flowId` (most relevant with `enqueue`), the flow's
        enablement + trust are re-validated (`enablementState ∈ {Enabled,
        UpdateAvailable}` ∧ `trustStatus ≠ untrusted`); a disabled / untrusted
        / pin-divergent flow is refused 422 `CONFIG` rather than stamped
        `triaged` + `auto` and left to stall the tick.
      operationId: extTriageTask
      parameters:
        - { name: slug, in: path, required: true, schema: { type: string } }
        - { name: taskId, in: path, required: true, schema: { type: string, format: uuid } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ExtTriageBody" }
            examples:
              verdict:
                summary: Verdict only — flow + runner, task becomes launchable
                value:
                  flowId: "bugfix"
                  runnerId: "claude-code"
                  targetBranch: "main"
                  promotionMode: "local_merge"
              flag-only:
                summary: (ADR-112) Flag for human review — no verdict fields
                value:
                  flag: true
              verdict-enqueue:
                summary: (ADR-112) Verdict + auto-enqueue after triage
                value:
                  flowId: "bugfix"
                  runnerId: "claude-code"
                  enqueue: true
      responses:
        "200":
          description: |
            Triage applied. `triageStatus` is `triaged` for a verdict body or
            `flagged` for a `flag: true` body.
          content:
            application/json:
              schema:
                type: object
                required: [ok, triageStatus]
                properties:
                  ok: { type: boolean }
                  triageStatus: { type: string, enum: [triaged, flagged] }
        "401": { $ref: "#/components/responses/TokenUnauthorized" }
        "403": { $ref: "#/components/responses/HitlForbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422":
          description: |
            Request body failed schema validation, or a `CONFIG` precondition
            was violated (ADR-112, Implemented) — `flag` sent with a verdict field
            (mutually exclusive), `enqueue` with no resolvable `flowId` (or with
            `flag`), or a verdict `flowId` that is disabled / untrusted /
            pin-divergent.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ExtErrorBody" }
...
components:
  securitySchemes:
    projectToken:
      type: http
      scheme: bearer
      description: |
        External API bearer token. Format: `mai_` + base64url(randomBytes(32)).
        Stored as sha256 hex at rest; never logged or streamed. Lookup is by
        the 12-character prefix; verification uses `timingSafeEqual`.
        Project-bound tokens carry `project_id` and must match the addressed
        project. Global personal user tokens carry `project_id = NULL`, require
        an active/password-clear owner, and authorize through that owner's
        current project access. A valid, active, non-expired, non-revoked token
        must hold the route's required scope or `*`, except exact
        `hitl:respond:human`, which `*` never implies. Missing scope returns
        403 without revealing held scopes. Token failures (invalid, expired,
        revoked, wrong-project, inaccessible-project) map to 401 or 404 per
        the existence-hide policy.
  responses:
    TokenUnauthorized:
      description: |
        Missing, invalid, expired, revoked, ownerless, inactive-owner, or
        password-change-required bearer token. Invalid prefix/hash mismatch,
        expired, revoked, and unusable-owner failures return 401.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ExtErrorBody" }
    NotFound:
      description: |
        Resource not found, or the token's project does not own the addressed
        resource. For global personal tokens, this also covers projects the
        owner cannot currently access. Existence-hide cross-project access
        returns 404, not 403.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ExtErrorBody" }
...
    HitlForbidden:
      description: |
        Access refused (`UNAUTHORIZED`). Common cases:

          - **Missing scope** — the token lacks the route's required scope and
            does not hold `*`. The body NEVER reveals which scopes the token
            holds.
          - **Exact human scope** — `hitl:respond:human` is required for
            `human`, `infra_recovery`, and `budget_breach` responses; `*`
            alone is insufficient.
          - **Actor-kind gate (D7, respond only)** — a project token,
            project-scoped user token, or agent token attempted to answer a
            human-only request. Those token actors may answer only
            `kind ∈ {permission, form}`.
          - **Wrong token kind (inbox only)** — `/api/v1/ext/hitl` requires a
            global personal user token; project/agent/project-scoped tokens
            return 403.

        Cross-project access does NOT use 403 — it hides existence (404).
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ExtErrorBody" }
  schemas:
    ExtErrorBody:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Machine-readable error code.
        message:
          type: string
          description: Human-readable error message.
    ExtTriageBody:
      type: object
      description: |
        (M34 — Implemented; ADR-112 `flag`/`enqueue` — Implemented) Triage body —
        at least one field. Every provided field is allow-list validated. A
        verdict body stamps `triage_status='triaged'`; a `flag: true` body
        stamps `triage_status='flagged'`.

        Cross-field rules (enforced server-side; OpenAPI 3.0 cannot express
        them declaratively), violations → 422 `CONFIG`:
          - `flag` is **mutually exclusive** with the verdict fields (`flowId`,
            `runnerId`, `baseBranch`, `targetBranch`, `promotionMode`).
          - `enqueue: true` is valid ONLY alongside a verdict that yields a
            `flowId` (this body's `flowId` or the task's existing
            `tasks.flow_id`); never with `flag`.
      minProperties: 1
      additionalProperties: false
      properties:
        flowId:
          type: string
          minLength: 1
          description: |
            Flow to run the task with; must belong to the project and (ADR-112,
            Implemented) be launchable (enabled + trusted + not pin-divergent).
        runnerId:
          type: string
          minLength: 1
          description: Verdict runner; must be an enabled catalog runner.
        baseBranch:
          type: string
          minLength: 1
          maxLength: 255
          description: Base branch the run worktree is created from (git ref-name shape).
        targetBranch:
          type: string
          minLength: 1
          maxLength: 255
          description: Promotion target branch (git ref-name shape).
        promotionMode:
          type: string
          enum: [local_merge, pull_request]
        flag:
          type: boolean
          description: |
            (ADR-112, Implemented) When `true`, mark the task `flagged` for human
            review (held — not launchable). Mutually exclusive with the verdict
            fields above; `flag: true` alone is a valid body.
        enqueue:
          type: boolean
          description: |
            (ADR-112, Implemented) When `true`, set the task to auto-launch after
            triage (`tasks.launch_mode = 'auto'`); the `auto_launch_triaged`
            tick performs the launch once the task is launchable. Valid only
            with a verdict that yields a `flowId`; never with `flag`.
...